Fix main.c and update led.c to correct driver
